+++ src/debug.rs
@@ -115,7 +115,7 @@ impl DebugContext {
 }
 
 impl Open for DebugContext {
-    fn open<App>(self, app: &mut App) -> crate::Result<Option<crate::window::WindowHandle>>
+    fn open<App>(self, app: &mut App) -> crate::Result<crate::window::WindowHandle>
     where
         App: Application + ?Sized,
     {